refactor: split recovery controller interface and implementation
