+++ aichallenge/workspace/src/aichallenge_submit/stuck_recovery_controller/CMakeLists.txt
@@ -5,7 +5,7 @@ find_package(ament_cmake_auto REQUIRED)
 ament_auto_find_build_dependencies()
 
 ament_auto_add_executable(stuck_recovery_controller_node
-  src/stuck_recovery_controller_node.cpp
+  src/stuck_recovery_controller.cpp
 )
 
 ament_auto_package()
